@@ -18,74 +18,83 @@
     "lint-fix": "eslint ./ --fix"
   },
   "dependencies": {
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@material-ui/styles": "^4.11.5",
-    "@monaco-editor/react": "^4.4.5",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@floating-ui/react-dom": "^2.1.0",
+    "@monaco-editor/react": "^4.6.0",
+    "@mui/icons-material": "^5.15.20",
+    "@mui/lab": "^5.0.0-alpha.170",
+    "@mui/material": "^5.15.20",
+    "@mui/styles": "^5.15.20",
     "@octokit/core": "^5.0.2",
     "@octokit/plugin-throttling": "^8.1.3",
-    "ansi_up": "^5.1.0",
+    "ansi_up": "^6.0.2",
     "atob": "^2.1.2",
     "aws-sdk": "^2.1223.0",
     "btoa": "^1.2.1",
-    "chalk": "^4.1.2",
-    "devicon": "^2.15.1",
-    "dotenv": "^10.0.0",
-    "highlight.js": "^11.5.1",
-    "iframe-resizer-react": "^1.1.0",
-    "katex": "^0.13.24",
-    "livecodes": "0.3.0",
-    "marked": "^4.0.17",
-    "next": "^12.1.0",
-    "next-i18next": "^8.10.0",
-    "next-query-params": "^2.2.1",
-    "nextjs-progressbar": "^0.0.13",
-    "node-fetch": "^3.2.4",
-    "notebookjs": "^0.6.7",
-    "ora": "<6.0.0",
-    "react": "^17.0.2",
-    "react-devicons": "^2.11.0-rev.2",
-    "react-dom": "^17.0.2",
-    "rehype-highlight": "<5.0.0",
-    "rehype-stringify": "<9.0.0",
-    "remark-gfm": "<2.0.0",
-    "remark-parse": "<10.0.0",
-    "remark-rehype": "<9.0.0",
-    "sanitize-html": "^2.12.1",
-    "title-case": "^3.0.3",
-    "typewriter-effect": "^2.19.0",
-    "unified": "<10.0.0",
-    "xterm-addon-fit": "^0.5.0",
+    "chalk": "^5.3.0",
+    "devicon": "^2.16.0",
+    "dotenv": "^16.4.5",
+    "highlight.js": "^11.9.0",
+    "iframe-resizer-react": "^1.0.13",
+    "katex": "^0.16.10",
+    "livecodes": "^0.5.0",
+    "marked": "^13.0.0",
+    "next": "^14.2.4",
+    "next-i18next": "^15.3.0",
+    "next-query-params": "^5.0.0",
+    "nextjs-progressbar": "^0.0.16",
+    "node-fetch": "^3.3.2",
+    "notebookjs": "^0.8.2",
+    "ora": "^8.0.1",
+    "react": "^18.3.1",
+    "react-devicons": "^2.16.2",
+    "react-dom": "^18.3.1",
+    "rehype-highlight": "^7.0.0",
+    "rehype-stringify": "^10.0.0",
+    "remark-gfm": "^4.0.0",
+    "remark-parse": "^11.0.0",
+    "remark-rehype": "^11.1.0",
+    "sanitize-html": "^2.13.0",
+    "stringify": "^2.2.0",
+    "title-case": "^4.3.1",
+    "typewriter-effect": "^2.21.0",
+    "unified": "^11.0.5",
+    "use-query-params": "^2.2.1",
+    "xterm-addon-fit": "^0.8.0",
     "xterm-for-react": "^1.0.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.0",
-    "@types/atob": "^2.1.2",
-    "@types/btoa": "^1.2.3",
-    "@types/cli-progress": "^3.11.0",
-    "@types/jsdom": "^16.2.14",
-    "@types/marked": "^4.0.3",
-    "@types/node": "^16.11.36",
-    "@types/node-fetch": "^3.0.3",
-    "@types/nprogress": "^0.2.0",
-    "@types/react": "^17.0.45",
-    "@types/sanitize-html": "^2.6.2",
-    "@typescript-eslint/eslint-plugin": "^5.25.0",
-    "@typescript-eslint/parser": "^5.25.0",
-    "babel-plugin-module-resolver": "^4.1.0",
-    "eslint": "<8.0.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-next": "^12.1.6",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-import-resolver-babel-module": "^5.3.1",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.30.0",
-    "eslint-plugin-react-hooks": "^4.5.0",
-    "prettier": "^2.6.2",
-    "ts-node": "^10.8.0",
-    "typescript": "^4.6.4"
-  }
+    "@babel/core": "^7.24.7",
+    "@types/atob": "^2.1.4",
+    "@types/btoa": "^1.2.5",
+    "@types/cli-progress": "^3.11.5",
+    "@types/jsdom": "^21.1.7",
+    "@types/marked": "^6.0.0",
+    "@types/node": "^20.14.8",
+    "@types/node-fetch": "^2.6.11",
+    "@types/nprogress": "^0.2.3",
+    "@types/react": "^18.3.3",
+    "@types/sanitize-html": "^2.11.0",
+    "@typescript-eslint/eslint-plugin": "^7.13.1",
+    "@typescript-eslint/parser": "^7.13.1",
+    "autoprefixer": "^10.4.19",
+    "babel-plugin-module-resolver": "^5.0.2",
+    "eslint": "^9.5.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-next": "^14.2.4",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-import-resolver-babel-module": "^5.3.2",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.9.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-react": "^7.34.3",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "postcss": "^8.4.38",
+    "prettier": "^3.3.2",
+    "tailwindcss": "^3.4.4",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.5.2"
+  },
+  "packageManager": "yarn@4.3.1"
 }
