@@ -1,4 +1,4 @@
-import { Button, Typography } from "@material-ui/core";
+import { Button, Typography } from "@mui/material";
 import NextLink from "next/link";
 import { CSSProperties } from "react";
 import classes from "./style.module.css";
