@@ -1,6 +1,6 @@
 import React from "react";
 import Document, { Html, Main, Head, NextScript } from "next/document";
-import { ServerStyleSheets } from "@material-ui/styles";
+import { ServerStyleSheets } from "@mui/styles";
 
 class MyDocument extends Document {
   render() {
