@@ -4,10 +4,10 @@ import { useRouter } from "next/router";
 import React, { useEffect, useMemo, useState } from "react";
 import dynamic from "next/dynamic";
 import { createNewPlayground } from "lib/playground";
-import { LinearProgress } from "@material-ui/core";
+import { LinearProgress } from "@mui/material";
 import { Algorithm } from "lib/models";
 import { Language } from "lib/repositories";
-import Alert from "@material-ui/lab/Alert";
+import Alert from "@mui/lab/Alert";
 import Head from "components/head";
 import PlaygroundLayout from "layouts/playground";
 import useTranslation from "hooks/translation";
